python_api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pillow==11.1.0
numpy==2.2.3
tensorflow>=2.13.0
tensorflow-datasets>=4.9.4
scikit-learn>=1.4.0
matplotlib>=3.8.0

